requirements.txt
PyGithub ==  1.55